31 add billing service implementation (#32)
* Add total billing

* Add billing

* Add billing rate service

* Swap to decimal

* Add implementation

* Update DI

* Update meter-api/meter-api/Services/BillingService.cs


* Update meter-api/meter-api/Services/BillingService.cs


---------

+++ meter-api/meter-api/Services/IBillingService.cs
@@ -4,6 +4,6 @@ namespace meter_api.Services
 {
     public interface IBillingService
     {
-        float CalculateCost(MeterAgentReading currentReading, MeterAgentReading? previousReading);
+        decimal CalculateCost(MeterAgentReading currentReading, MeterAgentReading? previousReading);
     }
 }
